requirements.txt
flask
flask-mysqldb
Flask-WTF
passlib
python-dotenv